# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Client Engagement Analyzer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface JobQueued {
  uploadId: string;
  status: string;
}

export type JobStatusStatus = typeof JobStatusStatus[keyof typeof JobStatusStatus];


export const JobStatusStatus = {
  pending: 'pending',
  processing: 'processing',
  done: 'done',
  error: 'error',
} as const;

export interface JobStatus {
  status: JobStatusStatus;
  step?: string | null;
  error?: string;
  startedAt: string;
}

export interface AnalyzeInput {
  uploadId: string;
}

export interface UploadResult {
  uploadId: string;
  filename: string;
  size: number;
  mimeType: string;
}

export interface TranscriptSegment {
  start: number;
  end: number;
  text: string;
  isQuestion: boolean;
  fillerWords: string[];
  /** Speaker ID (e.g. SPEAKER_00) — present when diarization ran */
  speaker?: string;
}

export interface TranscriptData {
  id: string;
  fullText: string;
  segments: TranscriptSegment[];
  questions: string[];
  fillerWords: string[];
}

export interface SpeakingMetrics {
  durationSeconds: number;
  totalWords: number;
  speakingRateWpm: number;
  questionsCount: number;
  fillerWordCount: number;
  fillerFrequencyPerMinute: number;
  pauseCount: number;
  avgPauseDurationSeconds: number;
  longestPauseDurationSeconds: number;
  energyMean: number;
  energyStd: number;
  pitchMean: number;
  pitchStd: number;
}

export type EngagementIndicatorClassification = typeof EngagementIndicatorClassification[keyof typeof EngagementIndicatorClassification];


export const EngagementIndicatorClassification = {
  Low: 'Low',
  Moderate: 'Moderate',
  High: 'High',
} as const;

export interface EngagementIndicator {
  score: number;
  classification: EngagementIndicatorClassification;
  explanation: string;
}

export interface EngagementIndicators {
  confidence: EngagementIndicator;
  engagement: EngagementIndicator;
  hesitation: EngagementIndicator;
  curiosity: EngagementIndicator;
  attentiveness: EngagementIndicator;
}

export interface ExecutiveSummary {
  overall: string;
  positiveObservations: string[];
  areasOfAttention: string[];
  recommendations: string[];
}

/**
 * Per-speaker communication metrics (present when diarization ran)
 */
export interface SpeakerMetrics {
  /** Human-readable label e.g. "Speaker 1" */
  label: string;
  speakingTimeSeconds: number;
  participationPct: number;
  wordCount: number;
  speakingRateWpm: number;
  questionsCount: number;
  fillerWordCount: number;
  fillerFrequencyPerMinute: number;
  indicators?: EngagementIndicators;
  summary?: ExecutiveSummary;
}

/**
 * Map of speaker ID to speaker profile (present when diarization ran)
 */
export type AnalysisResultSpeakers = {[key: string]: SpeakerMetrics};

export interface AnalysisResult {
  id: string;
  date: string;
  filename: string;
  hasDiarization?: boolean;
  transcript: TranscriptData;
  metrics: SpeakingMetrics;
  indicators: EngagementIndicators;
  summary: ExecutiveSummary;
  /** Map of speaker ID to speaker profile (present when diarization ran) */
  speakers?: AnalysisResultSpeakers;
  pdfAvailable: boolean;
}

export interface AnalysisSummary {
  id: string;
  date: string;
  filename: string;
  durationSeconds: number;
  overallEngagement: string;
  pdfAvailable: boolean;
  hasDiarization?: boolean;
  speakerCount?: number;
}