# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Client Engagement Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResultSpeakers } from './analysisResultSpeakers';
import type { EngagementIndicators } from './engagementIndicators';
import type { ExecutiveSummary } from './executiveSummary';
import type { SpeakingMetrics } from './speakingMetrics';
import type { TranscriptData } from './transcriptData';

export interface AnalysisResult {
  id: string;
  date: string;
  filename: string;
  hasDiarization?: boolean;
  transcript: TranscriptData;
  metrics: SpeakingMetrics;
  indicators: EngagementIndicators;
  summary: ExecutiveSummary;
  /** Map of speaker ID to speaker profile (present when diarization ran) */
  speakers?: AnalysisResultSpeakers;
  pdfAvailable: boolean;
}