# lib/api-zod/src/generated/types/analysisSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Client Engagement Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisSummary {
  id: string;
  date: string;
  filename: string;
  durationSeconds: number;
  overallEngagement: string;
  pdfAvailable: boolean;
  hasDiarization?: boolean;
  speakerCount?: number;
}