# lib/api-zod/src/generated/types/engagementIndicator.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Client Engagement Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { EngagementIndicatorClassification } from './engagementIndicatorClassification';

export interface EngagementIndicator {
  score: number;
  classification: EngagementIndicatorClassification;
  explanation: string;
}