# lib/api-zod/src/generated/types/engagementIndicators.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Client Engagement Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { EngagementIndicator } from './engagementIndicator';

export interface EngagementIndicators {
  confidence: EngagementIndicator;
  engagement: EngagementIndicator;
  hesitation: EngagementIndicator;
  curiosity: EngagementIndicator;
  attentiveness: EngagementIndicator;
}