# lib/api-zod/src/generated/types/jobQueued.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Client Engagement Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface JobQueued {
  uploadId: string;
  status: string;
}