# lib/api-zod/src/generated/types/jobStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Client Engagement Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatusStatus } from './jobStatusStatus';

export interface JobStatus {
  status: JobStatusStatus;
  step?: string | null;
  error?: string;
  startedAt: string;
}