# lib/api-zod/src/generated/types/speakingMetrics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Client Engagement Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface SpeakingMetrics {
  durationSeconds: number;
  totalWords: number;
  speakingRateWpm: number;
  questionsCount: number;
  fillerWordCount: number;
  fillerFrequencyPerMinute: number;
  pauseCount: number;
  avgPauseDurationSeconds: number;
  longestPauseDurationSeconds: number;
  energyMean: number;
  energyStd: number;
  pitchMean: number;
  pitchStd: number;
}