# lib/api-zod/src/generated/types/transcriptData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Client Engagement Analyzer API
 * OpenAPI spec version: 0.1.0
 */
import type { TranscriptSegment } from './transcriptSegment';

export interface TranscriptData {
  id: string;
  fullText: string;
  segments: TranscriptSegment[];
  questions: string[];
  fillerWords: string[];
}