# lib/api-zod/src/generated/types/transcriptSegment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Client Engagement Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface TranscriptSegment {
  start: number;
  end: number;
  text: string;
  isQuestion: boolean;
  fillerWords: string[];
  /** Speaker ID (e.g. SPEAKER_00) — present when diarization ran */
  speaker?: string;
}