# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Client Engagement Analyzer API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResult {
  uploadId: string;
  filename: string;
  size: number;
  mimeType: string;
}